core/use-cases/register-to-event.ts
import { z } from "zod";
import { getKnexClient } from "../lib/knex";
import { AttendeesRepository } from "../repositories/attendees-repository";
import { RegistrationsRepository } from "../repositories/registrations-repository";

export async function registerToEvent(
  command: RegistrationCommand
): Promise<void> {
  const satinitizedFirstName = satinizeFirstName(command.firstName);
  validateFirstName(satinitizedFirstName);

  await registerAttendee({
    firstName: satinitizedFirstName,
    lastName: command.lastName,
  });
}

function satinizeFirstName(firstName: string) {
  return firstName.trim();
}

function validateFirstName(firstName: string) {
  const firstNameRegex =
    /^[a-zA-ZàáâäãåąćęèéêëėįìíîïłńòóôöõøùúûüųūÿýżźñçčšžÀÁÂÄÃÅĄĆČĖĘÈÉÊËÌÍÎÏĻŁŃÒÓÔÖÕØÙÚÛÜŲŪŸÝŻŹÑßÇŒÆŠŽ∂ð .'-]+$/u;

  const firstNameSchema = z.object({
    firstName: z.string().min(2).max(250).regex(firstNameRegex),
  });

  firstNameSchema.parse({ firstName });
}

async function registerAttendee(command: RegistrationCommand) {
  const knexClient = await getKnexClient();

  const attendeesRepository = new AttendeesRepository(knexClient);
  const attendeeId = await attendeesRepository.add({
    first_name: command.firstName,
    last_name: command.lastName,
  });

  const registerRepository = new RegistrationsRepository(knexClient);
  await registerRepository.add(1, attendeeId);
}

export interface RegistrationCommand {
  lastName: string;
  firstName: string;
}
